backend/user-service/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
Flask-CORS
Flask-Migrate
Flask-Session